+++ src/lang/modules.ml
@@ -50,7 +50,6 @@ let () =
       "synth";
       "synth.all";
       "video";
-      "video.add_text";
       "video.external";
       "video.frame";
       "video.testsrc";